Rename simulate_spe->horace_simulate_spe test func
During experimentation trying to solve the problem of tests failing when
test_sqw was being run concurrently on two processes, it was found that
simulate_spe_testfunc existed in Horace and Herbert.

I was attempting to alter the config to solve the above problem, which
involved calling herbert_config.init_tests. This added
Herbert/_test/common_functions to the path ABOVE
Horace/_test/common_functions - which meant the tests now called
Herbert's version of simulate_spe_testfunc which was causing an error.
Prefixing the function with "horace_" avoids any naming ambiguity
between Herbert and Horace version of the script.


Former-commit-id: 318dbb591b455da4fd4a065f7e12f263c6a97a79