Add horace_set_local_parallel_config function
This function moves the config directory to somewhere local so that, if
tests are running at the same time in two different processes, they are
not attempting to read/write the same config files. I think this is what
was happening on the Windows build server, where the 2019b and 2018b
jobs were running concurrently, and test_sqw was failing.


Former-commit-id: ffd6c051e1baf1ad34747d99a8098b172463ccc0