Make small optimisation to compute_bin_data_mex_
As part of the algorithm, there is "find" call that searches for the
first positive number in an array. It does this on every iteration. On
every iteration a scalar is subtracted from the array, so the index
returned by "find" cannot increase. This means the "find" only needs to
search from the index that was found on the previous iteration


Former-commit-id: 99efd3d3c424c4637fbf862ad53067c9bf91daa3